Fix 64bit to 32bit demotions.
Many parts of the code allowed implicit conversion from a 64 bit value
to a 32 bit value.  A lot of this was using IceTSizeType as a 32 bit
integer.  Using the compiler warnings, remove the conversions or make
them explicit.

+++ tests/CompressionSize.c
@@ -27,7 +27,7 @@ static IceTDouble IdentityMatrix[16] = {
     0.0, 0.0, 1.0, 0.0,
     0.0, 0.0, 0.0, 1.0
 };
-static IceTFloat Black[4] = { 0.0, 0.0, 0.0, 1.0 };
+static IceTFloat Black[4] = { 0.0f, 0.0f, 0.0f, 1.0f };
 
 static void InitPathologicalImage(IceTImage image)
 {
@@ -81,7 +81,7 @@ static void InitActiveImage(IceTImage image)
     IceTSizeType num_pixels;
     int seed;
 
-    seed = time(NULL);
+    seed = (int)time(NULL);
     srand(seed);
 
     num_pixels = icetImageGetNumPixels(image);
